# api.yaml
openapi: 3.0.3
info:
  title: Hospital Management System API
  description: >
    REST API for a hospital management system supporting three roles —
    Admin, Doctor, and Patient — with JWT-based authentication, appointment
    scheduling, treatment records, AI-powered patient summaries, and a
    post-appointment chat feature.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Unauthorized

    # ─── Doctor ──────────────────────────────────────────────────────────────
...
    Patient:
      type: object
      properties:
        patient_id:
          type: integer
        patient_name:
          type: string
        patient_username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        gender:
          type: string
        age:
          type: integer
        address:
          type: string
        is_blacklisted:
          type: boolean
...
    Appointment:
      type: object
      properties:
        app_id:
          type: integer
        patient_id:
          type: integer
        doctor_id:
          type: integer
        date:
          type: string
          format: date
        time:
          type: string
          format: time
          example: "09:00"
        status:
          type: string
          enum: [BOOKED, COMPLETED, CANCELLED]
        visit_type:
          type: string
          enum: [Morning, Afternoon, Evening]
        token_no:
          type: integer
          nullable: true
          description: Queue position within the slot (patient dashboard only)

    # ─── Treatment ───────────────────────────────────────────────────────────
    Treatment:
      type: object
      properties:
        treatment_id:
          type: integer
        app_id:
          type: integer
        diagnosis:
          type: string
        prescription:
          type: string
        medication:
          type: string
        tests_done:
          type: string
        next_visit_date:
          type: string
          format: date
          nullable: true
    TreatmentRequest:
      type: object
      properties:
        diagnosis:
          type: string
        prescription:
          type: string
        medication:
          type: string
        tests_done:
          type: string
        next_visit_date:
          type: string
          format: date
          nullable: true

    # ─── Chat ────────────────────────────────────────────────────────────────
...
    AvailabilitySetRequest:
      type: object
      description: >
        Keys are ISO dates (YYYY-MM-DD) for the next 7 days.
        Values are arrays of slot names to mark as available.
      example:
        availability:
          "2025-07-01": ["Morning", "Evening"]
          "2025-07-02": ["Afternoon"]
      properties:
        availability:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
              enum: [Morning, Afternoon, Evening]

    # ─── Auth ────────────────────────────────────────────────────────────────
...
paths:
...
  /api/doctor/patients/{patient_id}:
    get:
      tags: [Doctor]
      summary: View a patient's full record and appointment history
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient profile with appointment history
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          patient:
                            $ref: '#/components/schemas/Patient'
                          appointments:
                            type: array
                            items:
                              $ref: '#/components/schemas/Appointment'
  /api/doctor/patients/{patient_id}/blacklist:
    post:
      tags: [Doctor]
      summary: Blacklist a patient (doctor action)
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient blacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /api/doctor/patients/{patient_id}/unblacklist:
    post:
      tags: [Doctor]
      summary: Remove blacklist from a patient (doctor action)
      parameters:
        - in: path
          name: patient_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Patient unblacklisted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/doctor/appointments/{app_id}/complete:
    post:
      tags: [Doctor]
      summary: Mark a BOOKED appointment as COMPLETED
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "400":
          description: Appointment is not in BOOKED status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Appointment belongs to a different doctor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/doctor/appointments/{app_id}/cancel:
    post:
      tags: [Doctor]
      summary: Cancel a BOOKED appointment
      parameters:
        - in: path
          name: app_id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        "400":
          description: Appointment already completed or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/doctor/appointments/{app_id}/treatment:
    parameters:
      - in: path
        name: app_id
        required: true
        schema:
          type: integer
    get:
      tags: [Doctor]
      summary: Retrieve treatment details for an appointment
      responses:
        "200":
          description: Appointment and treatment data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          appointment:
                            $ref: '#/components/schemas/Appointment'
                          treatment:
                            $ref: '#/components/schemas/Treatment'
    post:
      tags: [Doctor]
      summary: Create or update treatment details for an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TreatmentRequest'
      responses:
        "200":
          description: Treatment saved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Treatment'
...
  /api/doctor/availability:
    get:
      tags: [Doctor]
      summary: Get the doctor's availability for the next 7 days
      responses:
        "200":
          description: Slots map for the next 7 days
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          dates:
                            type: array
                            items:
                              type: string
                              format: date
                          slots:
                            type: object
                            description: Slot definitions with start/end times
                          slots_map:
                            type: object
                            description: >
                              Nested object — date → slot → boolean indicating availability
    post:
      tags: [Doctor]
      summary: Set or replace availability for the next 7 days
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilitySetRequest'
      responses:
        "200":
          description: Availability updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...